+++ dubbo-annotation-processor/src/main/java/org/apache/dubbo/annotation/util/ASTUtils.java
@@ -1,0 +1,73 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.dubbo.annotation.util;
+
+import org.apache.dubbo.annotation.AnnotationProcessorContext;
+
+import com.sun.source.util.TreePath;
+import com.sun.tools.javac.code.Symbol;
+import com.sun.tools.javac.tree.JCTree;
+import com.sun.tools.javac.util.ListBuffer;
+
+import java.util.List;
+
+/**
+ * Some utils about AST manipulating.
+ */
+public final class ASTUtils {
+
+    private ASTUtils() {
+        throw new UnsupportedOperationException("No instance of 'ASTUtils' for you! ");
+    }
+
+    public static void addImportStatement(AnnotationProcessorContext apContext,
+                                    Symbol.ClassSymbol classSymbol,
+                                    String packageName,
+                                    String className) {
+
+        JCTree.JCImport jcImport = apContext.getTreeMaker().Import(
+            apContext.getTreeMaker().Select(
+                apContext.getTreeMaker().Ident(apContext.getNames().fromString(packageName)),
+                apContext.getNames().fromString(className)
+            ), false);
+
+        TreePath treePath = apContext.getTrees().getPath(classSymbol);
+        TreePath parentPath = treePath.getParentPath();
+        JCTree.JCCompilationUnit compilationUnit = (JCTree.JCCompilationUnit) parentPath.getCompilationUnit();
+
+        List<JCTree.JCImport> imports = compilationUnit.getImports();
+        if (imports.stream().noneMatch(x -> x.qualid.toString().contains(packageName + "." + className))) {
+
+            compilationUnit.accept(new JCTree.Visitor() {
+                @Override
+                public void visitTopLevel(JCTree.JCCompilationUnit that) {
+
+                    List<JCTree> defs = compilationUnit.defs;
+
+                    ListBuffer<JCTree> newDefs = new ListBuffer<>();
+
+                    newDefs.add(defs.get(0));
+                    newDefs.add(jcImport);
+                    newDefs.addAll(defs.subList(1, defs.size()));
+
+                    compilationUnit.defs = newDefs.toList();
+                }
+            });
+        }
+    }
+}
